refactor: update Layout
